requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.28.2
selenium>=4.9.0
webdriver-manager>=3.8.6
sqlalchemy>=2.0.9
python-multipart>=0.0.6
beautifulsoup4>=4.12.2
lxml>=4.9.2
markdown>=3.4.3
pymongo>=4.3.3
aiofiles>=23.1.0
httpx>=0.24.0
jinja2>=3.1.2
pytest>=7.3.1
pytest-asyncio>=0.21.0
debugpy